Remove the old SingleParticipantFragment, changes new of tabbed/body to single/group
