+++ app/src/main/scala/com/waz/zclient/participants/ParticipantsController.scala
@@ -100,8 +100,7 @@ class ParticipantsController(implicit injector: Injector, context: Context, ec: 
   def withCurrentConvGroupOrBot(callback: Callback[java.lang.Boolean]): Unit =
     isGroupOrBot.head.foreach { groupOrBot => callback.callback(groupOrBot) }(Threading.Ui)
 
-
-  def showRemoveConfirmation(userId: UserId) = getUser(userId).foreach {
+  def showRemoveConfirmation(userId: UserId): Unit = getUser(userId).foreach {
     case Some(userData) =>
       val request = new ConfirmationRequest.Builder()
         .withHeader(getString(R.string.confirmation_menu__header))
@@ -122,6 +121,6 @@ class ParticipantsController(implicit injector: Injector, context: Context, ec: 
       confirmationController.requestConfirmation(request, IConfirmationController.PARTICIPANTS)
       inject[SoundController].playAlert()
     case _ =>
-  }
+  }(Threading.Ui)
 
 }
